refactor: rearrange project structure
