+++ src/styles/searchImages.css
